requirements.txt
ipython
numpy
pandas
matplotlib
scikit-learn
seaborn
google-cloud-storage
pyyaml
imbalanced-learn
mlflow
dvc
flask

#-e .